+++ src/prefect/runner/__init__.py
@@ -1,4 +1,5 @@
+from .aio import AsyncRunner
 from .runner import Runner
 from .submit import submit_to_runner, wait_for_submitted_runs
 
-__all__ = ["Runner", "submit_to_runner", "wait_for_submitted_runs"]
+__all__ = ["AsyncRunner", "Runner", "submit_to_runner", "wait_for_submitted_runs"]
